+++ src/buffer.c
@@ -37,18 +37,18 @@ void DeleteBuffer(StringBuffer *Buffer){
 }
 
 void ZeroBuffer(StringBuffer *Buffer){
-    for(uint32_t i = 0; i < Buffer->Length; i++)
+    for(int i = 0; i < Buffer->Length; i++)
         Buffer->Memory[i] = '\0';
 }
 
-void ZeroBufferEx(StringBuffer *Buffer, uint32_t StartIndex){
-    for(uint32_t i = StartIndex; i < Buffer->Length; i++)
+void ZeroBufferEx(StringBuffer *Buffer, int StartIndex){
+    for(int i = StartIndex; i < Buffer->Length; i++)
         Buffer->Memory[i] = '\0';
 }
 
 void AppendBuffer(StringBuffer *Buffer, char *Str){
-    uint32_t BufferStringLength = StringLength(Buffer->Memory);
-    uint32_t StringSize = StringLength(Str);
+    int BufferStringLength = StringLength(Buffer->Memory);
+    int StringSize = StringLength(Str);
 
     while(BufferStringLength + StringSize > Buffer->Length)
         DoubleSize(Buffer);
@@ -56,16 +56,16 @@ void AppendBuffer(StringBuffer *Buffer, char *Str){
     StringConcat(Buffer->Memory, Str);
 }
 
-uint32_t AppendBufferEx(StringBuffer *Buffer, char *Str, uint32_t MaxLength, uint32_t Offset){
-    uint32_t BufferStringLength = StringLength(Buffer->Memory);
-    uint32_t StringSize = StringLength(Str);
+uint32_t AppendBufferEx(StringBuffer *Buffer, char *Str, int MaxLength, int Offset){
+    int BufferStringLength = StringLength(Buffer->Memory);
+    int StringSize = StringLength(Str);
 
-    uint32_t ActualSize = SmallerUnsigned(StringSize, MaxLength);
+    int ActualSize = SmallerUnsigned(StringSize, MaxLength);
 
     while(BufferStringLength + ActualSize > Buffer->Length)
         DoubleSize(Buffer);
 
-    MemoryCopy((Buffer->Memory) + BufferStringLength - (uint32_t)1, Str + Offset, ActualSize);
+    MemoryCopy((Buffer->Memory) + BufferStringLength - 1, Str + Offset, ActualSize);
     *((Buffer->Memory) + BufferStringLength + ActualSize - 1) = '\0';
 
     return BufferStringLength + ActualSize;
@@ -73,7 +73,7 @@ uint32_t AppendBufferEx(StringBuffer *Buffer, char *Str, uint32_t MaxLength, uin
 
 // StringBufferArray
 
-StringBufferArray CreateBufferArray(uint32_t InitialMaxElements){
+StringBufferArray CreateBufferArray(int InitialMaxElements){
     StringBufferArray temp;
     temp.NumberOfElements = 0;
     temp.MaxNumberOfElements = InitialMaxElements;
@@ -81,14 +81,14 @@ StringBufferArray CreateBufferArray(uint32_t InitialMaxElements){
 
     temp.Data = HeapAlloc(hHeap, 0, temp.MemorySize);
 
-    for(uint32_t i = 0; i < InitialMaxElements; i++)
+    for(int i = 0; i < InitialMaxElements; i++)
         *StringBufferGetElemenetAt(&temp, i) = CreateBuffer(64);
 
     return temp;
 }
 
 void DoubleArrayCapacity(StringBufferArray *Array){
-    uint32_t OldCapacity = Array->MaxNumberOfElements;
+    int OldCapacity = Array->MaxNumberOfElements;
     Array->MaxNumberOfElements *= 2;
     Array->MemorySize = sizeof(StringBuffer) * Array->MaxNumberOfElements;
 
@@ -99,12 +99,12 @@ void DoubleArrayCapacity(StringBufferArray *Array){
     Array->Data = temp;
 
 
-    for(uint32_t i = OldCapacity; i < Array->MaxNumberOfElements; i++)
+    for(int i = OldCapacity; i < Array->MaxNumberOfElements; i++)
         *StringBufferGetElemenetAt(Array, i) = CreateBuffer(64);
 }
 
 void DeleteBufferArray(StringBufferArray *Array){
-    for (uint32_t i = 0; i < Array->NumberOfElements; i++) {
+    for (int i = 0; i < Array->NumberOfElements; i++) {
         StringBuffer *target = StringBufferGetElemenetAt(Array, i);
         HeapFree(hHeap, 0, target->Memory);
     }
@@ -117,7 +117,7 @@ void DeleteBufferArray(StringBufferArray *Array){
     Array->NumberOfElements = 0;
 }
 
-StringBuffer *StringBufferGetElemenetAt(StringBufferArray *Array, uint32_t ElementIndex){
+StringBuffer *StringBufferGetElemenetAt(StringBufferArray *Array, int ElementIndex){
     
     if(ElementIndex >= Array->MaxNumberOfElements)
         return NULL;
@@ -125,11 +125,11 @@ StringBuffer *StringBufferGetElemenetAt(StringBufferArray *Array, uint32_t Eleme
     return &(Array->Data[ElementIndex]);
 }
 
-void InsertLine(StringBufferArray *Array, uint32_t ElementIndex, char *Str){
+void InsertLine(StringBufferArray *Array, int ElementIndex, char *Str){
 
     while(Array->NumberOfElements + 1 >= Array->MaxNumberOfElements) DoubleArrayCapacity(Array);
 
-    for(uint32_t i = Array->NumberOfElements; i > ElementIndex; i--) {
+    for(int i = Array->NumberOfElements; i > ElementIndex; i--) {
         /*StringBuffer *dest = StringBufferGetElemenetAt(Array, i + 1);
         StringBuffer *src = StringBufferGetElemenetAt(Array, i);
         *dest = *src;*/
@@ -148,22 +148,22 @@ void InsertLine(StringBufferArray *Array, uint32_t ElementIndex, char *Str){
     StringCopy(target->Memory, Str);
 }
 
-void RemoveLine(StringBufferArray *Array, uint32_t ElementIndex){
+void RemoveLine(StringBufferArray *Array, int ElementIndex){
     
     DeleteBuffer(StringBufferGetElemenetAt(Array, ElementIndex));
 
-    for(uint32_t i = ElementIndex; i < Array->NumberOfElements; i++)
+    for(int i = ElementIndex; i < Array->NumberOfElements; i++)
         *StringBufferGetElemenetAt(Array, i) = *StringBufferGetElemenetAt(Array, i + 1);
 
 }
 
-void RemoveLineEx(StringBufferArray *Array, uint32_t ElementIndex, char *OutStr){
+void RemoveLineEx(StringBufferArray *Array, int ElementIndex, char *OutStr){
     if(OutStr)
         StringCopy(OutStr, StringBufferGetElemenetAt(Array, ElementIndex)->Memory);
     
     DeleteBuffer(StringBufferGetElemenetAt(Array, ElementIndex));
 
-    for(uint32_t i = ElementIndex; i < Array->NumberOfElements - 1; i++)
+    for(int i = ElementIndex; i < Array->NumberOfElements - 1; i++)
         *StringBufferGetElemenetAt(Array, i) = *StringBufferGetElemenetAt(Array, i + 1);
 
     StringBuffer *last = StringBufferGetElemenetAt(Array, Array->NumberOfElements - 1);
@@ -171,9 +171,9 @@ void RemoveLineEx(StringBufferArray *Array, uint32_t ElementIndex, char *OutStr)
     last->Length = 0;
 }
 
-uint32_t MaxLineLength(StringBufferArray *Array){
-    uint32_t MaxLength = 0;
-    for(uint32_t i = 0; i < Array->NumberOfElements; i++){
+int MaxLineLength(StringBufferArray *Array){
+    int MaxLength = 0;
+    for(int i = 0; i < Array->NumberOfElements; i++){
         StringBuffer *temp = StringBufferGetElemenetAt(Array, i);
         if(StringLength(temp->Memory)> MaxLength) MaxLength = StringLength(temp->Memory);
     }
